Moved all blosxom_* related items to its own file
With over 1300 lines of code, pyblosxom is getting ridiculously large
and difficult to maintain.

Moved all the blosxom_* utility functions to its own module, and turns
out, Pyblosxom only needs two of those functions.

This will lead to easier refactorings later. For example
`blosxom_process_path_info` should probably be an object, and could be a
candidate for Change Method into Method Object refactoring.

`blosxom_hander` is another candidate.

pyblosxom.py also has its warts, `Pyblosxom.run_static_renderer` is too
long and needs to be cleaned up.

This patch should pave the way for later updates.

Changes:
`blosxom.py` - new file containing everything `blosxom_*`
`pyblosxom.py` - removed all `blosxom_*`
`plugins/tags.py` - this plugin requires the `blosxom_*` stuff
`plugins/xmlrpc_pingback.py` - Actually, should be removed, requires
functions not available anymore in comments.py. What say you?
`tests/*` - updated blosxom related tests.
